--- frontend/nginx.conf
@@ -1,7 +1,0 @@
-server {
-  listen  80;
-  
-  location /auth {
-      root /usr/share/nginx/html;
-  }
-}
