+++ frontend/package.json
@@ -1,23 +1,20 @@
 {
-  "name": "frontend",
+  "name": "test-react-admin",
   "version": "0.1.0",
-  "private": true,
   "homepage": "/auth",
+  "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/material": "^5.2.7",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^17.0.2",
+    "notistack": "^2.0.5",
+    "prop-types": "^15.8.1",
+    "ra-data-json-server": "^4.2.0",
+    "react": "^18.2.0",
     "react-admin": "^4.2.0",
-    "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "web-vitals": "^2.1.3"
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
